Add macOS target and other fixes
- Make designer default on VS Mac again
- Remove cruft
- Fix loading issues on VS for Windows by providing the binding path
