+++ src/Eto.DevExtension.VisualStudio.Mac/Eto.DevExtension.VisualStudio.Mac.csproj
@@ -11,7 +11,6 @@
   </PropertyGroup>
   
   <ItemGroup>
-    <Compile Remove="Editor\XetoTextEditorExtension.cs" />
     <Compile Include="..\Eto.DevExtension.VisualStudio.Windows\Intellisense\*.cs" Link="Intellisense\%(Filename)%(Extension)" />
     <Compile Include="..\Eto.DevExtension.VisualStudio.Windows\Components.cs" />
   </ItemGroup>
