+++ src/Eto.DevExtension.VisualStudio.Mac/Properties/Manifest.addin.xml
@@ -36,7 +36,7 @@
 			imageId="eto-project"
 			wizard="Eto.DevExtension.VisualStudio.Mac.ProjectWizard"
 			category="multiplat/app/eto"
-			supportedParameters="AppName,Combined,Xeto,Preview,Jeto,XamMac,Framework"
+			supportedParameters="AppName,Combined,Xeto,Preview,Jeto,XamMac,macos,Framework"
 			/>
 		<Template
 			id="Eto.App.FSharp"
@@ -44,7 +44,7 @@
 			imageId="eto-project"
 			wizard="Eto.DevExtension.VisualStudio.Mac.ProjectWizard"
 			category="multiplat/app/eto"
-			supportedParameters="AppName,Combined,Xeto,Preview,Jeto,XamMac,Framework"
+			supportedParameters="AppName,Combined,Xeto,Preview,Jeto,XamMac,macos,Framework"
 			/>
 		<Template
 			id="Eto.App.VisualBasic"
@@ -52,7 +52,7 @@
 			imageId="eto-project"
 			wizard="Eto.DevExtension.VisualStudio.Mac.ProjectWizard"
 			category="multiplat/app/eto"
-			supportedParameters="AppName,Combined,Xeto,Preview,Jeto,XamMac,Framework"
+			supportedParameters="AppName,Combined,Xeto,Preview,Jeto,XamMac,macos,Framework"
 			/>
 	</Extension>
 
@@ -118,14 +118,6 @@
 		<!--ProjectFlavor guid="" type="Eto.DevExtension.VisualStudio.Mac.FSharpProjectExtension" alias="Eto.FSharp"/-->
 	</Extension>	
 	
-	<!-- MD 7
-	<Extension path="/MonoDevelop/Ide/DisplayBindings">
-		<DisplayBinding insertbefore="DefaultDisplayBinding" class="Eto.DevExtension.VisualStudio.Mac.Editor.DisplayBinding" />
-	</Extension>
-	-->
-	<!-- <Extension path="/MonoDevelop/Ide/TextEditorExtensions">
-		<Class fileExtensions=".xeto" class="Eto.DevExtension.VisualStudio.Mac.Editor.XetoTextEditorExtension" insertbefore="XmlTextEditorExtension" />
-	</Extension> -->
 	<Extension path="/MonoDevelop/Core/MimeTypes">
 		<MimeType id="application/x-xeto" _description="Eto Xaml document" icon="md-xml-file-icon" isText="true" baseType="application/xml">
 			<File pattern="*.xeto" />
