+++ src/Eto.DevExtension.VisualStudio.Windows/Templates/App-CSharp/App.vstemplate
@@ -31,7 +31,7 @@
 			<CustomParameter Name="$uistyle$" Value="none"/>
 			<CustomParameter Name="$groupid$" Value="Eto.App" />
 			<CustomParameter Name="SideWaffleNewProjNode" Value="Project\CSharp\Eto.Forms"/>
-			<CustomParameter Name="SupportedParameters" Value="Combined,Xeto,Jeto,Preview,XamMac,Framework" />
+			<CustomParameter Name="SupportedParameters" Value="Combined,Xeto,Jeto,Preview,XamMac,macos,Framework" />
 		</CustomParameters>
 	</TemplateContent>
 	<WizardExtension>
